refactor: have separated test-utils in web-components and web-marketplace
